agrego eliminar pub de alquiler

+++ YateMate/Amarra/views.py
@@ -55,8 +55,12 @@ def mis_publicaciones(request):
 
 def eliminar_publicacion(request, id):
     publicacion = get_object_or_404(Publicacion_Amarra, id=id)
-    publicacion.delete()
-    messages.success(request, 'Publicacion eliminada')
+    if Reserva.objects.filter(publicacion=publicacion).exists:
+        messages.success(request, "Esta operación no es posible, existe una reserva") 
+        return redirect ('mis_publicaciones')
+    else:    
+        publicacion.delete()
+        messages.success(request, 'Publicacion eliminada')
     return redirect('list_amarra')
 
 def modificar_publicacion(request, id):
